requirements.txt
openenv-core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
gradio>=4.20.0
plotly>=5.20.0
transformers>=4.40.0
torch>=2.2.0
pytest>=7.4.0
huggingface_hub>=0.20.0
unsloth